Activation Keys and Active
Added functionality in which user can add keys that should activate the object. Pressing these keys while the object is active will call the key input function.
Added functionality to turn objects or containers off. An inactive object will only run its frame update function to be able to turn back on and will therefore not be drawn to the screen or call any of the objects it contains.