Bug Fix when Clicking on Rotated Object
When moving the mouse an inexact distance away from the target object (scroll bar), the object would forget it is being clicked and revert its position.
The code to rotate the mouse position by the objects rotation has been replaced by vector math as it is more efficient and works better.